requirements.txt
selenium==4.39.0
Django==6.0.1
djangorestframework==3.16.1